create common and fix T2
