+++ mozhegova.maria/common/delimiter.cpp
@@ -1,0 +1,17 @@
+#include "delimiter.hpp"
+
+std::istream & mozhegova::operator>>(std::istream & in, DelimiterIO && dest)
+{
+  std::istream::sentry sentry(in);
+  if (!sentry)
+  {
+    return in;
+  }
+  char c = '0';
+  in >> c;
+  if (in && (c != dest.exp))
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
