+++ mozhegova.maria/common/scopeGuard.cpp
