+++ mozhegova.maria/common/scopeGuard.hpp
